requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
bs4==0.0.1
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
drf-yasg==1.21.5
enum-compat==0.0.3
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
newrelic==8.7.1
packaging==23.0
psycopg2==2.9.5
pyTelegramBotAPI==4.10.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.1.0
sentry-sdk==1.17.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
telebot==0.0.5
tqdm==4.65.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
yarl==1.8.2
